# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Документация API для управления продуктами и категориями",
        "version": "1.0.0"
    },
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить список всех продуктов",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Список продуктов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Создать новый продукт",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Продукт успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка создания продукта",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Продукт не был создан"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка создания"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить продукт по ID",
                "operationId": "b46f4aa73232d0368eee698a082342a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID продукта",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные продукта",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Продукт не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Продукт не был найден"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка получения продукта"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Обновить продукт по ID",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID продукта",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Продукт успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Продукт не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Продукт не был найден"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка обновления продукта"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Непредвиденная ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Непредвиденная ошибка"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Удалить продукт по ID",
                "operationId": "fe34b364a3f192305598b65d04d06b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID продукта",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Продукт успешно удален"
                    },
                    "404": {
                        "description": "Продукт не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Продукт не найден"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка удаления продукта"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Непредвиденная ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Непредвиденная ошибка"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}/category": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить категорию продукта",
                "operationId": "1cb838f4c7800367dbc953a37286b649",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID продукта",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория продукта",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Продукт не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Продукт не найден"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка получения категории продукта"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/category/{category_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Фильтрация продуктов по категории",
                "operationId": "989b12b5909c8802be5ea26f647befdf",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Продукты, отфильтрованные по категории",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/price-range": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Фильтрация продуктов по диапазону цен",
                "operationId": "7cf23bf6fc049678487160d16de4160a",
                "parameters": [
                    {
                        "name": "minPrice",
                        "in": "query",
                        "description": "Минимальная цена",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "maxPrice",
                        "in": "query",
                        "description": "Максимальная цена",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Продукты, отфильтрованные по диапазону цен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/sort-by-price": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Сортировка продуктов по цене",
                "operationId": "2757ff77a89f764e3cac7ab2225e4836",
                "parameters": [
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Порядок сортировки: asc (по возрастанию) или desc (по убыванию)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Отсортированные продукты",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/search": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Поиск продукта по имени",
                "operationId": "3c3311837393a14e95bd05609b231141",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Название продукта для поиска",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Продукты, найденные по имени",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreProductRequest": {
                "title": "Store Product Request",
                "description": "Запрос на создание нового продукта",
                "required": [
                    "name",
                    "price",
                    "category_id"
                ],
                "properties": {
                    "name": {
                        "description": "Название продукта",
                        "type": "string",
                        "example": "Товар A"
                    },
                    "description": {
                        "description": "Описание продукта",
                        "type": "string",
                        "example": "Описание товара A"
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "number",
                        "format": "float",
                        "example": 199
                    },
                    "category_id": {
                        "description": "ID категории продукта",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateProductRequest": {
                "title": "Update Product Request",
                "description": "Запрос на обновление продукта",
                "properties": {
                    "name": {
                        "description": "Название продукта",
                        "type": "string",
                        "example": "Товар A"
                    },
                    "description": {
                        "description": "Описание продукта",
                        "type": "string",
                        "example": "Новое описание товара A"
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "number",
                        "format": "float",
                        "example": 249
                    },
                    "category_id": {
                        "description": "ID категории продукта",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Схема для категории",
                "properties": {
                    "id": {
                        "description": "ID категории",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название категории",
                        "type": "string"
                    },
                    "description": {
                        "description": "Описание категории",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "description": "Схема для продукта",
                "properties": {
                    "id": {
                        "description": "ID продукта",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название продукта",
                        "type": "string"
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "number",
                        "format": "float"
                    },
                    "category_id": {
                        "description": "ID категории",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
...
}